+++ skgstat/__init__.py
@@ -5,5 +5,5 @@ from .Kriging import OrdinaryKriging
 from . import interfaces
 
 # set some stuff
-__version__ = '0.2.7'
+__version__ = '0.2.8'
 __author__ = 'Mirko Maelicke <mirko.maelicke@kit.edu>'