+++ skgstat/interfaces/__init__.py
@@ -1,1 +1,3 @@
 from .variogram_estimator import VariogramEstimator
+from .pykrige import pykrige_model, pykrige_params, pykrige_as_kwargs
+from .gstools import gstools_cov_model